Retrieve accession numbers and titles for a given combination of species and genes in genbank

Retrieve accession numbers and titles for complex searches conducted in genbank. Note that this function depends on acc.retrieve as it actually uses expand.grid to find all the relevant organism/gene combinations in genbank.

acc.table.retrieve(clades, species, genes, speciesLevel)

Arguments

clades
A vector of clade names (character). Note that this can either be the name of a clade (e.g. Apis) or the code in NCBI (e.g. txid7459).

species
A vector of species names (logical). Note that this can either be the name of a clade (e.g. Apis) or the code in NCBI (e.g. txid7459).

genes
A vector of gene names (character; optional).

speciesLevel
Whether the result should be a species-level dataset (logical).

Value

data.frame

Examples

if (FALSE) {
acc.table.retrieve(
 clades  = c('Felis', 'Vulpes', 'Phoca'),
 species = 'Manis_pentadactyla' ,
 genes   = c("A2AB","ADORA3","ADRB2","APOB",
            "APP","ATP7","BCHE","BDNF",
            "BMI1","BRCA1","BRCA2","CNR1",
            "COI","CREM","CYTB","DMP1",
            "EDG1","ENAM","FBN1","GHR",
            "IRBP","ND1","ND2","PLCB4",
            "PNOC","RAG1a","RAG1b","RAG2",
            "TTN","TYR1","VWF"),
 speciesLevel=FALSE
)
}